`dask`, more structure for nbs and format tweaks
Added in an image directory for notebooks just to tidy things up, and
spent a lot of time getting to know how to use dask (see
`data/nbs/what-is-dask.ipynb` and issue #5 for details).

From there, the formatting of the datasets has been tweaked to utilise dask's dataframe collection for manipulation (and eventual collation). As such, `environment.yml` has been updated also.