Update formatting, found missing rows.
Slight restructure of the formatting functions. This finalises the
preprocessing of the data (hopefully) and as a bonus the missing half a
million entries have been recovered. The dataset itself is smaller by
capacity (down from some 2.4GB to 1.99GB) though this is likely due to
the removal of some other columns (?) and specifying dtypes. Perhaps.
Not too concerned so long as it's all there.