Quasi bipartite graphs (#50)
* Add notebook for quasi-bipartite sampling

* Start writing scripts for sampler

* Finish initial draft for sampling script

Sampling is done by taking a seeded random sample from each sub-frame's
rows. Then adjacency matrices and dissimilarity matrices are made
lazily. The resulting similar pair dataframes are made lazily as well.
All of these functions are then brought in and a collection of the
resulting dataframes are concatenated into one dask dataframe.

* Write worker writing function

* Finish first edit of QBG sampler; needs improving

Although implemented with dask, is it actually working as well as it
could?

Lots more work needed on how to determine splits and inputs effectively.

* Update .gitignore

* Update notebooks

* Update .gitignore to remove results